First and last patterns do not clash
Previously it was possible for the first generated pattern to be the same as the last generated pattern for a given color list. When generating the last pattern in a list the first and also previous pattern are excluded. This improves readability in charts where the first and last category are contiguous. Updates #6.